Pantum M7300FDW workaround revisited
Previous approach relaxed IPP parsing rules everywhere.

This version allows IPP violation only on responses and (as much as
possible) only when we speak with hardware printer, not with another
instance of CUPS.

Default IPP behavior in libcups is not affected in comparison with the
original version, so tools are not affected, including the IPP Everywhere™
Printer Self-Certification Tools.