Changed the end-of-search criteria for ippfind -T 0
Previously, with the -T 0 option ("search quickly with automatic
timeout"), ippfind used the following criteria to determine when the
search was complete:

  1. At least one service was discovered, and
  2. Either all discovered services were resolved or 2.5 seconds had
     passed since the last update.

However, this algorithm caused unreliable behavior when a mix of devices
was present—some that could be discovered quickly (such as cached or
locally published devices via ipp-usb) and others that required more
time (such as network devices). As a result, ippfind often terminated
prematurely, missing network devices in its output.

The updated end-of-search criteria are as follows:

  1. All discovered services must be either resolved or have at least 1.0
      second since their last update, and
  2. The total discovery time of at least 2.5 seconds has passed.

This updated algorithm significantly improves discovery reliability.